ScheduleController Policy
Role & Permission doctor's schedule

+++ app/Http/Controllers/Doctors/ScheduleController.php
@@ -6,6 +6,7 @@ use App\Models\Doctor\Schedule;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\ScheduleRequest;
+use App\Models\Appointment\Appointment;
 
 class ScheduleController extends Controller
 {
@@ -29,9 +30,9 @@ class ScheduleController extends Controller
     public function store(ScheduleRequest $request)
     {
         $checkSchedules = Schedule::where([
-            ['doctor_id', Auth::user()->doctor->id],
-            ['day', $request->get('day')],
-        ]);
+                            ['doctor_id', Auth::user()->doctor->id],
+                            ['day', $request->get('day')],
+                        ]);
 
         if ($checkSchedules->first()) {
             flash('Warning! Schedule you selected is already exist')->warning();
@@ -39,7 +40,7 @@ class ScheduleController extends Controller
             return redirect()->back();
         }
 
-        $schedule = Schedule::create($request->formSchedule());
+        Schedule::create($request->formSchedule());
 
         flash('Successful! New schedule created')->success();
 
@@ -48,6 +49,8 @@ class ScheduleController extends Controller
 
     public function show(Schedule $schedule)
     {
+        $this->authorize('update', $schedule);
+
         $appointments = Appointment::with('patient')->where('schedule_id', $schedule->id)->get();
 
         return view('people.schedule.show', compact('schedule', 'appointments'));
@@ -55,11 +58,15 @@ class ScheduleController extends Controller
 
     public function edit(Schedule $schedule)
     {
+        $this->authorize('update', $schedule);
+
         return view('people.schedule.edit', compact('schedule'));
     }
 
     public function update(ScheduleRequest $request, Schedule $schedule)
     {
+        $this->authorize('update', $schedule);
+
         $schedule->fill($request->formSchedule());
         $schedule->save();
 
@@ -70,6 +77,8 @@ class ScheduleController extends Controller
 
     public function destroy(Schedule $schedule)
     {
+        $this->authorize('update', $schedule);
+
         $schedule->delete();
 
         flash('Successful! Schedule deleted.')->success();
